# backend/static/openapi.yaml
openapi: "3.0.3"
info:
  title: "IOT Smart Home for Elderly — Alfred API"
  description: |
    REST API cho hệ thống nhà thông minh hỗ trợ người cao tuổi.
    
    ## Xác thực
    Tất cả endpoint (trừ `/api/auth/login`, `/api/auth/register`) đều yêu cầu signed bearer token.
    Token được ký bằng `itsdangerous` và được gửi qua **HttpOnly Cookie** (`batman_os_auth`) hoặc header `Authorization: Bearer <token>`.
    
    ## Rate Limiting
    - Login: 5 req/phút, 20 req/giờ
    - AI Chat: 30 req/phút, 300 req/giờ
    - Devices: 60 req/phút, 2000 req/giờ
    - Mặc định: 100 req/giờ, 500 req/ngày
  version: "2.0.0"
  contact:
    name: "Tu Dinh Huy — ITITIU21213"
    email: "ITITIU21213@student.hcmiu.edu.vn"
  license:
    name: "Academic (Thesis Project)"
servers:
  - url: "http://127.0.0.1:5000"
    description: "Local development"
  - url: "https://your-domain.com"
    description: "Production"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: batman_os_auth
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          example: "heart_rate"
        message:
          type: string
          example: "Nhịp tim bất thường: 130 BPM"
        severity:
          type: string
          enum: [normal, caution, warning, critical]
          example: "warning"
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time
        device_code:
          type: string
          nullable: true
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/alerts:
    get:
      tags: [alerts]
      summary: Lấy danh sách cảnh báo
      parameters:
        - in: query
          name: severity
          schema:
            type: string
            enum: [normal, caution, warning, critical]
        - in: query
          name: is_read
          schema:
            type: boolean
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Danh sách cảnh báo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Alert"
                  total:
                    type: integer
  /api/alerts/sos:
    post:
      tags: [alerts]
      summary: Kích hoạt SOS khẩn cấp
      description: Gửi email SOS đến người thân/caretaker, tùy chọn đính kèm voice note
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                note:
                  type: string
                  maxLength: 200
                  example: "Tôi bị ngã, cần giúp đỡ!"
                audio:
                  type: string
                  format: binary
                  description: Voice note (webm/wav/mp3, tối đa 5MB)
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  maxLength: 200
      responses:
        "200":
          description: SOS đã gửi
        "429":
          description: Rate limited (10 req/phút)
...